Search for the associated conanfile
The path to the conanfile.py is needed to call conan install on it.
Until now this path was hardcoded for the first try-outs.
By checking on the same level and one level above the current project
file the conanfile.py is searched. If one is found the path is used,
otherwise aborted.