Introduce manual set conanfile path
The automatic detection is based on a specific file organization, e.g. the
conanfile must be stored on the same level or one level above the file
hirarchy. To allow the user the freedom to decide where the conanfile
is placed the ini configuration mechanism is added. The additional
qconan.ini is always searched for on the project level and if present
this file can be used to reference the conanfile relative to this
directory.