Initial Commit
Everything starts with an idea, also this project. The project already started in Christmas 2019 with a
few tryouts on QtCreator to get a feeling about the possibility. Along with some tryouts,
this readme.md was created to collect first ideas about the project and to find a way
how to start with the project.