Fix conan install run on QtCreator exit
If a CMake project is used and the QtCreator application is
quit, conan install is executed again because the currentProjectChanged
signal is emitted on the same project.
The used build directory from the last conan install command is
stored and compared again the next conan install call. This way a
second call with the same build directory can be detected and prohibit.
In case the conanfile is changed conan install must be called on the
same build directory.