Use conan build information to access shared libraries
Windows shared libraries can only be found if they are accessible within the PATH
environment variable or in the same directory. To allow usage of conan packages
during development, this means use a DLL located in the conan cache, the path to
the conan package must be added to the PATH environment variable.
This commit uses the build information created during conan install to collect
all library and binary folders referenced by the required packages. This collection
is added as environment PATH information to the active project.