Use environment PATH from conan build information
Every conan package can manipulate the PATH variable in the package_info
method. The consuming conan package combine this information and expose it
in the build information.
This commit use this information by forwarding the provided information to
the project. The library and binary folder information can be
used as environment PATH if the associated configuration flag,
useLibPath or useBinPath, is set to true.